# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Reservation API",
        "description": "API para gestionar reservas de espacios en eventos",
        "contact": {
            "email": "soporte@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/spaces": {
            "get": {
                "summary": "Obtener espacios disponibles",
                "operationId": "e17708e9e7050fff0a684603393f619b",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Tipo de espacio (auditorium, meeting_room, conference_room)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "capacity",
                        "in": "query",
                        "description": "Capacidad mínima del espacio",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Fecha para comprobar disponibilidad",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de espacios disponibles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Crear un nuevo espacio",
                "operationId": "8bbbe73fe6d28fc7aadf84f49791569b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "capacity",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Sala de reuniones"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Una sala perfecta para reuniones"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "meeting_room"
                                    },
                                    "image": {
                                        "description": "Archivo de imagen opcional",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Espacio creado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "space": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/spaces/{id}": {
            "get": {
                "summary": "Obtener un espacio por ID",
                "operationId": "42a320419e7c2f110e1cf34132f9e26a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del espacio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Espacio obtenido con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "space": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Espacio no encontrado"
                    }
                }
            },
            "put": {
                "summary": "Actualizar un espacio",
                "operationId": "ad0300609c73b94572ef9b932a2fc6db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del espacio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "capacity",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Sala de conferencias"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Sala para grandes conferencias"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "conference_room"
                                    },
                                    "image": {
                                        "description": "Archivo de imagen opcional",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Espacio actualizado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "space": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Espacio no encontrado"
                    }
                }
            },
            "delete": {
                "summary": "Eliminar un espacio",
                "operationId": "5a5eda39c31c16b6218dc227d58898a1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del espacio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Espacio eliminado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Space deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Espacio no encontrado"
                    }
                }
            }
        }
    }
}